Don't delete blobs - they may be used in several other objects. Implement the Blob Garbage Collector instead...

+++ Zetbox.API.Common/Packaging/Importer.cs
@@ -126,6 +126,9 @@ namespace Zetbox.App.Packaging
                 Log.InfoFormat("Deleting {0} objects marked for deletion", objectsToDelete.Count());
                 foreach (var pairToDelete in objectsToDelete)
                 {
+                    // Don't delete blobs, the blob garbage collector should delete them.
+                    if (pairToDelete.Value is Blob) continue; 
+
                     ctx.Delete(pairToDelete.Value);
                 }
                 Log.Info("Deployment finished");
